+++ src/main/java/ru/ziplla/repository/VacancyRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.ziplla.entity.Vacancy;
 
 @Repository
-public interface VacancyRepository extends JpaRepository<Vacancy, Integer> {
+public interface VacancyRepository extends JpaRepository<Vacancy, Long> {
 }
